Assistant checkpoint: Add admin and public interfaces with search functionality
Assistant generated file changes:
- app.py: Add admin routes and public search
- templates/public/search.html: Create public search page
- templates/public/search_results.html: Create search results page
- templates/public/product_detail.html: Create public product detail page

---

User prompt:

This is working great! Now, I would like to make a minor change to the way things are on the user front end. I would like to only be able to access everything we just built by going to an admin url or a hidden url that no one would guess. I would then like for the front page of the app to look like a search engine for products/batch numbers that when a pproduct or batch number is serched the user is presented with a similar page as we have on the admin side, but they WILL NOT have PDF generation functionlity or access to the PDF label that was generated. Of course, since they are just a viewer, they will also not have access tro edit buttons and delete buttons. as well. In the place where the Generated PDF is on the admin side, there will be a large download icon/link that they can click to download a PDF of a COA for the selected product. The url on the front end for the particulat product will be {base_url}/{batch_number}

Can you do this?