Assistant checkpoint: Update template feature to support optional attribute values
Assistant generated file changes:
- models.py: Update ProductTemplate model to store attribute values
- templates/template_create.html: Update template creation form to support values
- static/js/template.js: Update template.js to handle attribute values

---

User prompt:

I would like to build on to the attribute templates feature. We need to be able to define values for some of the attribute name. Not all of them though, so filling in the input for the value should be options. But, we should be able to pre fill attribute names and vlaues