Assistant checkpoint: Add Square sync endpoint and UI integration
Assistant generated file changes:
- square_sync.py: Add single product sync endpoint, Update sync function with location ID
- templates/product_detail.html: Update sync button handler

---

User prompt:

Now, in the @square_sync.pycan we start trying to test the square inntegration by tying it to the Sync to Square button on the product details page? Let's try not to mess with any other code unless we absolutely have to. Also, remeber, there is already a square_catalog_id in the product table. Let's start nice and slow. We also need to be able to see the responses and errors so we know what to fix. We will also need a location ID, but since this will be static, we can add this as a secret.