Assistant checkpoint: Add Square update reminder modal
Assistant generated file changes:
- templates/product_detail.html: Add Square update reminder modal, Add JavaScript for Square reminder modal

---

User prompt:

I have an idea; if there is a square id attached and we update the product, I would like a modal to popup when we are sent back to the details page that will remind us that this is an item in square and not to forget to update the item in square. Then we will have the update in square button there as well as a cancel button that will close the modal. Make it a pretty modal like the rest of the others ones on the site.