Assistant checkpoint: Add settings page with development toggles
Assistant generated file changes:
- models.py: Add Settings model
- templates/settings.html: Create settings page template
- app.py: Add settings routes
- templates/base.html: Add settings link to navbar
- templates/product_edit.html: Add settings check for Square ID controls

---

User prompt:

Can you add a settings page. On this settings page let have a section with a header that says "Development Setting". And undet this there will be toggles that will allow me to toggle on and off the Clear Square Id and Image Id functionlity. When off they will not be present in the edit product page. When on they will be. That way we won't have to delete these from the code and we can use them later on for tesing, but turn them off when we are not testing.