Assistant checkpoint: Add Square delete confirmation modal
Assistant generated file changes:
- templates/product_detail.html: Add Square delete confirmation modal, Update unsync button click handler, Add Square delete confirmation modal, Update unsync button click handler

---

User prompt:

Can you add a modal popup for the delete from square that is similar to the delete product popup confirmation so we can make sure that the person meant to click the delete button? Do this without removing the functionality of any of the other buttons. Just add this functionlity and leave everything else in place.